+++ sender.go
@@ -300,8 +300,6 @@ func (s *Sender) mux() {
 		close(s.l.done)
 	}()
 
-	var clientClosed bool // indicates a client-side close
-
 Loop:
 	for {
 		var outgoingTransfers chan frames.PerformTransfer
@@ -313,7 +311,7 @@ Loop:
 		}
 
 		closed := s.l.close
-		if clientClosed {
+		if s.l.closeInProgress {
 			// swap out channel so it no longer triggers
 			closed = nil
 		}
@@ -329,8 +327,12 @@ Loop:
 			// populated queue
 			fr := *q.Dequeue()
 			s.l.rxQ.Release(q)
-			s.l.doneErr = s.muxHandleFrame(fr, clientClosed)
-			if s.l.doneErr != nil {
+
+			// if muxHandleFrame returns an error it means the mux must terminate.
+			// note that in the case of a client-side close due to an error, nil
+			// is returned in order to keep the mux running to ack the detach frame.
+			if err := s.muxHandleFrame(fr); err != nil {
+				s.l.doneErr = err
 				return
 			}
 
@@ -353,13 +355,16 @@ Loop:
 				continue Loop
 			}
 
-		case err := <-closed:
-			// sender is being closed by the client (Close() or protocol error)
-			clientClosed = true
+		case <-closed:
+			if s.l.closeInProgress {
+				// a client-side close due to protocol error is in progress
+				continue
+			}
+			// sender is being closed by the client
+			s.l.closeInProgress = true
 			fr := &frames.PerformDetach{
 				Handle: s.l.handle,
 				Closed: true,
-				Error:  err,
 			}
 			_ = s.l.session.txFrame(fr, nil)
 
@@ -373,7 +378,7 @@ Loop:
 
 // muxHandleFrame processes fr based on type.
 // depending on the peer's RSM, it might return a disposition frame for sending
-func (s *Sender) muxHandleFrame(fr frames.FrameBody, clientClosed bool) error {
+func (s *Sender) muxHandleFrame(fr frames.FrameBody) error {
 	debug.Log(2, "RX (Sender): %s", fr)
 	switch fr := fr.(type) {
 	// flow control frame
@@ -410,7 +415,7 @@ func (s *Sender) muxHandleFrame(fr frames.FrameBody, clientClosed bool) error {
 		case s.l.session.tx <- resp:
 			debug.Log(2, "TX (Sender): %s", resp)
 		case <-s.l.close:
-			return &LinkError{}
+			return nil
 		case <-s.l.session.done:
 			return s.l.session.doneErr
 		}
@@ -434,7 +439,7 @@ func (s *Sender) muxHandleFrame(fr frames.FrameBody, clientClosed bool) error {
 		case s.l.session.tx <- dr:
 			debug.Log(2, "TX (Sender): mux frame to Session: %d, %s", s.l.session.channel, dr)
 		case <-s.l.close:
-			return &LinkError{}
+			return nil
 		case <-s.l.session.done:
 			return s.l.session.doneErr
 		}
@@ -442,7 +447,7 @@ func (s *Sender) muxHandleFrame(fr frames.FrameBody, clientClosed bool) error {
 		return nil
 
 	default:
-		return s.l.muxHandleFrame(fr, clientClosed)
+		return s.l.muxHandleFrame(fr)
 	}
 
 	return nil